+++ homeassistant/components/geniushub/__init__.py
@@ -8,8 +8,6 @@ from geniushubclient import GeniusHub
 import voluptuous as vol
 
 from homeassistant.const import (
-    ATTR_ENTITY_ID,
-    ATTR_TEMPERATURE,
     CONF_HOST,
     CONF_MAC,
     CONF_PASSWORD,
@@ -17,7 +15,6 @@ from homeassistant.const import (
     CONF_USERNAME,
     TEMP_CELSIUS,
 )
-from homeassistant.core import callback
 from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.discovery import async_load_platform
@@ -27,7 +24,6 @@ from homeassistant.helpers.dispatcher import (
 )
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.event import async_track_time_interval
-from homeassistant.helpers.service import verify_domain_control
 from homeassistant.helpers.typing import ConfigType, HomeAssistantType
 import homeassistant.util.dt as dt_util
 
@@ -71,28 +67,6 @@ CONFIG_SCHEMA = vol.Schema(
 ATTR_ZONE_MODE = "mode"
 ATTR_DURATION = "duration"
 
-SVC_SET_ZONE_MODE = "set_zone_mode"
-SVC_SET_ZONE_OVERRIDE = "set_zone_override"
-
-SET_ZONE_MODE_SCHEMA = vol.Schema(
-    {
-        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
-        vol.Required(ATTR_ZONE_MODE): vol.In(["off", "timer", "footprint"]),
-    }
-)
-SET_ZONE_OVERRIDE_SCHEMA = vol.Schema(
-    {
-        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
-        vol.Required(ATTR_TEMPERATURE): vol.All(
-            vol.Coerce(float), vol.Range(min=4, max=28)
-        ),
-        vol.Optional(ATTR_DURATION): vol.All(
-            cv.time_period,
-            vol.Range(min=timedelta(minutes=5), max=timedelta(days=1)),
-        ),
-    }
-)
-
 
 async def async_setup(hass: HomeAssistantType, config: ConfigType) -> bool:
     """Create a Genius Hub system."""
@@ -121,70 +95,7 @@ async def async_setup(hass: HomeAssistantType, config: ConfigType) -> bool:
     for platform in ["climate", "water_heater", "sensor", "binary_sensor", "switch"]:
         hass.async_create_task(async_load_platform(hass, platform, DOMAIN, {}, config))
 
-    setup_service_functions(hass, broker)
-
     return True
-
-
-@callback
-def setup_service_functions(hass: HomeAssistantType, broker):
-    """Set up the service functions."""
-
-    @verify_domain_control(hass, DOMAIN)
-    async def set_zone_mode(call) -> None:
-        """Set the system mode."""
-        entity_id = call.data[ATTR_ENTITY_ID]
-
-        registry = await hass.helpers.entity_registry.async_get_registry()
-        registry_entry = registry.async_get(entity_id)
-
-        if registry_entry is None or registry_entry.platform != DOMAIN:
-            raise ValueError(f"'{entity_id}' is not a known {DOMAIN} entity")
-
-        if registry_entry.domain != "climate":
-            raise ValueError(f"'{entity_id}' is not an {DOMAIN} zone")
-
-        payload = {
-            "unique_id": registry_entry.unique_id,
-            "service": call.service,
-            "data": call.data,
-        }
-
-        async_dispatcher_send(hass, DOMAIN, payload)
-
-    async def set_switch_override(call) -> None:
-        """Set the system mode."""
-        entity_id = call.data[ATTR_ENTITY_ID]
-
-        registry = await hass.helpers.entity_registry.async_get_registry()
-        registry_entry = registry.async_get(entity_id)
-
-        if registry_entry is None or registry_entry.platform != DOMAIN:
-            raise ValueError(f"'{entity_id}' is not a known {DOMAIN} entity")
-
-        if registry_entry.domain != "switch":
-            raise ValueError(f"'{entity_id}' is not an {DOMAIN} zone")
-
-        payload = {
-            "unique_id": registry_entry.unique_id,
-            "service": call.service,
-            "data": call.data,
-        }
-        async_dispatcher_send(hass, DOMAIN, payload)
-
-    hass.services.async_register(
-        DOMAIN,
-        SVC_SET_SWITCH_OVERRIDE,
-        set_switch_override,
-        schema=SET_SWITCH_OVERRIDE_SCHEMA,
-    )
-
-    hass.services.async_register(
-        DOMAIN, SVC_SET_ZONE_MODE, set_zone_mode, schema=SET_ZONE_MODE_SCHEMA
-    )
-    hass.services.async_register(
-        DOMAIN, SVC_SET_ZONE_OVERRIDE, set_zone_mode, schema=SET_ZONE_OVERRIDE_SCHEMA
-    )
 
 
 class GeniusBroker:
@@ -313,32 +224,6 @@ class GeniusZone(GeniusEntity):
         self._zone = zone
         self._unique_id = f"{broker.hub_uid}_zone_{zone.id}"
 
-    async def _refresh(self, payload: Optional[dict] = None) -> None:
-        """Process any signals."""
-        if payload is None:
-            self.async_schedule_update_ha_state(force_refresh=True)
-            return
-
-        if payload["unique_id"] != self._unique_id:
-            return
-
-        if payload["service"] == SVC_SET_ZONE_OVERRIDE:
-            temperature = round(payload["data"][ATTR_TEMPERATURE] * 10) / 10
-            duration = payload["data"].get(ATTR_DURATION, timedelta(hours=1))
-
-            await self._zone.set_override(temperature, int(duration.total_seconds()))
-            return
-
-        mode = payload["data"][ATTR_ZONE_MODE]
-
-        # pylint: disable=protected-access
-        if mode == "footprint" and not self._zone._has_pir:
-            raise TypeError(
-                f"'{self.entity_id}' can not support footprint mode (it has no PIR)"
-            )
-
-        await self._zone.set_mode(mode)
-
     @property
     def name(self) -> str:
         """Return the name of the climate device."""
@@ -389,9 +274,3 @@ class GeniusHeatingZone(GeniusZone):
     def supported_features(self) -> int:
         """Return the bitmask of supported features."""
         return self._supported_features
-
-    async def async_set_temperature(self, **kwargs) -> None:
-        """Set a new target temperature for this zone."""
-        await self._zone.set_override(
-            kwargs[ATTR_TEMPERATURE], kwargs.get(ATTR_DURATION, 3600)
-        )
