+++ homeassistant/components/geniushub/climate.py
@@ -1,5 +1,8 @@
 """Support for Genius Hub climate devices."""
+from datetime import timedelta
 from typing import List, Optional
+
+import voluptuous as vol
 
 from homeassistant.components.climate import ClimateEntity
 from homeassistant.components.climate.const import (
@@ -13,6 +16,8 @@ from homeassistant.components.climate.const import (
     SUPPORT_PRESET_MODE,
     SUPPORT_TARGET_TEMPERATURE,
 )
+from homeassistant.const import ATTR_ENTITY_ID, ATTR_TEMPERATURE
+from homeassistant.helpers import config_validation as cv, entity_platform
 from homeassistant.helpers.typing import ConfigType, HomeAssistantType
 
 from . import DOMAIN, GeniusHeatingZone
@@ -25,6 +30,31 @@ HA_PRESET_TO_GH = {PRESET_ACTIVITY: "footprint", PRESET_BOOST: "override"}
 GH_PRESET_TO_HA = {v: k for k, v in HA_PRESET_TO_GH.items()}
 
 GH_ZONES = ["radiator", "wet underfloor"]
+
+ATTR_ZONE_MODE = "mode"
+ATTR_DURATION = "duration"
+
+SVC_SET_ZONE_MODE = "set_zone_mode"
+SVC_SET_ZONE_OVERRIDE = "set_zone_override"
+
+SET_ZONE_MODE_SCHEMA = vol.Schema(
+    {
+        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
+        vol.Required(ATTR_ZONE_MODE): vol.In(["off", "timer", "footprint"]),
+    }
+)
+SET_ZONE_OVERRIDE_SCHEMA = vol.Schema(
+    {
+        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
+        vol.Required(ATTR_TEMPERATURE): vol.All(
+            vol.Coerce(float), vol.Range(min=4, max=28)
+        ),
+        vol.Optional(ATTR_DURATION): vol.All(
+            cv.time_period,
+            vol.Range(min=timedelta(minutes=5), max=timedelta(days=1)),
+        ),
+    }
+)
 
 
 async def async_setup_platform(
@@ -42,6 +72,21 @@ async def async_setup_platform(
             for z in broker.client.zone_objs
             if z.data["type"] in GH_ZONES
         ]
+    )
+
+    # Register custom services
+    platform = entity_platform.current_platform.get()
+
+    platform.async_register_entity_service(
+        SVC_SET_ZONE_MODE,
+        SET_ZONE_MODE_SCHEMA,
+        "async_set_zone",
+    )
+
+    platform.async_register_entity_service(
+        SVC_SET_ZONE_OVERRIDE,
+        SET_ZONE_OVERRIDE_SCHEMA,
+        "async_set_temperature",
     )
 
 
@@ -101,3 +146,17 @@ class GeniusClimateZone(GeniusHeatingZone, ClimateEntity):
     async def async_set_preset_mode(self, preset_mode: str) -> None:
         """Set a new preset mode."""
         await self._zone.set_mode(HA_PRESET_TO_GH.get(preset_mode, "timer"))
+
+    async def async_set_zone(self, **kwargs) -> None:
+        """Set a new mode for this zone."""
+        mode = kwargs.get(ATTR_ZONE_MODE)
+
+        if mode == "footprint" and not self._zone._has_pir:
+            raise TypeError(
+                f"'{self.entity_id}' can not support footprint mode (it has no PIR)"
+            )
+        await self._zone.set_mode(mode)
+
+    async def async_set_temperature(self, **kwargs) -> None:
+        """Set a new target temperature for and options duration this zone."""
+        await self._zone.set_override(kwargs[ATTR_TEMPERATURE], int(kwargs.get(ATTR_DURATION, 3600).total_seconds()))
